% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_attraction.R
\name{add_attraction}
\alias{add_attraction}
\title{Add attraction of cells using RNA velocity}
\usage{
add_attraction(dataset)
}
\arguments{
\item{dataset}{A dataset created by \code{\link[=wrap_data]{wrap_data()}} or \code{\link[=wrap_expression]{wrap_expression()}}}
}
\value{
A dynwrap object with the attraction added.
}
\description{
Add attraction of cells using RNA velocity
}
