% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_cell_waypoints.R
\name{add_cell_waypoints}
\alias{add_cell_waypoints}
\alias{is_wrapper_with_waypoint_cells}
\title{Add cell waypoints to a wrapped trajectory}
\usage{
add_cell_waypoints(traj, num_cells_selected = 100)

is_wrapper_with_waypoint_cells(traj)
}
\arguments{
\item{traj}{Trajectory object}

\item{num_cells_selected}{About the number of cells selected as waypoints}
}
\description{
Add cell waypoints to a wrapped trajectory
}
\keyword{adapt_trajectory}
