% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_cyclic_trajectory_to_wrapper.R
\name{add_cyclic_trajectory_to_wrapper}
\alias{add_cyclic_trajectory_to_wrapper}
\title{Add a cyclic trajectory to a data wrapper}
\usage{
add_cyclic_trajectory_to_wrapper(data_wrapper, pseudotimes, ...)
}
\arguments{
\item{data_wrapper}{A data wrapper to extend upon.}

\item{pseudotimes}{A named vector of pseudo times.}

\item{...}{extra information to be stored in the wrapper.}
}
\description{
This function will generate the milestone_network and progressions.
}
