% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_dimred_projection_to_wrapper.R
\name{add_dimred_projection_to_wrapper}
\alias{add_dimred_projection_to_wrapper}
\title{Add a dimred projection trajectory to a data wrapper}
\usage{
add_dimred_projection_to_wrapper(data_wrapper, milestone_network,
  dimred_milestones, dimred_cells, milestone_assignment_cells = NULL,
  num_segments_per_edge = 100, ...)
}
\arguments{
\item{data_wrapper}{A data wrapper to extend upon.}

\item{milestone_network}{A network of milestones.}

\item{dimred_milestones}{A dimensionality reduction of the milestones.}

\item{dimred_cells}{A dimensionality reduction of the cells.}

\item{milestone_assignment_cells}{A milestone assignment of the cells.}

\item{num_segments_per_edge}{The number of pieces each milestone transition will be split info for the calculation of the percentages.}

\item{...}{extra information to be stored in the wrapper.}
}
\description{
This function will generate the milestone_network and progressions.
}
