% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_end_state_probabilities.R
\name{add_end_state_probabilities}
\alias{add_end_state_probabilities}
\title{Multifurcating trajectory with end state probabilities}
\usage{
add_end_state_probabilities(model, end_state_probabilities,
  pseudotime = NULL, do_scale_minmax = TRUE, ...)
}
\arguments{
\item{model}{The model to which a multifurcating trajectory will be added.}

\item{end_state_probabilities}{A dataframe containing cell_id (character) and additional numeric columns containing the probability for every end milestone. If the data_frame contains only a cell_id column, the data will be processed using \code{add_linear_trajectory}}

\item{pseudotime}{A named vector of pseudo times.}

\item{do_scale_minmax}{Whether or not to scale the pseudotime between 0 and 1.
Otherwise, will assume the values are already within that range.}

\item{...}{Extras to be added to the model}
}
\value{
The trajectory model
}
\description{
Constructs a multifurcating trajectory using the pseudotime values of each cell and their end state probabilities.
If pseudotime values are not given, will use pseudotime already present in the model.
}
\details{
This function will generate the milestone_network and progressions.
}
\keyword{create_trajectory}
