% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_trajectory.R
\name{add_trajectory}
\alias{add_trajectory}
\alias{is_wrapper_with_trajectory}
\title{Define a trajectory model given its milestone network and milestone percentages or progressions}
\usage{
add_trajectory(model, milestone_ids = NULL, milestone_network,
  divergence_regions = NULL, milestone_percentages = NULL,
  progressions = NULL, allow_self_loops = FALSE, ...)

is_wrapper_with_trajectory(model)
}
\arguments{
\item{model}{The model to which the trajectory will be added.}

\item{milestone_ids}{The ids of the milestones in the trajectory. Type: Character vector.}

\item{milestone_network}{The network of the milestones.
Type: Data frame(from = character, to = character, length = numeric, directed = logical).}

\item{divergence_regions}{A data frame specifying the divergence
regions between milestones (e.g. a bifurcation).
Type: Data frame(divergence_id = character, milestone_id = character, is_start = logical).}

\item{milestone_percentages}{A data frame specifying what percentage milestone each cell
consists of.
Type: Data frame(cell_id = character, milestone_id = character, percentage = numeric).}

\item{progressions}{Specifies the progression of a cell along a transition in the milestone_network.
Type: Data frame(cell_id = character, from = character, to = character, percentage = numeric).}

\item{allow_self_loops}{Whether to allow self loops
Type: Logical}

\item{...}{extra information to be stored in the model}
}
\value{
The trajectory model
}
\description{
Define a trajectory model given its milestone network and milestone percentages or progressions
}
