% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allowed_inputs}
\alias{allowed_inputs}
\title{All allowed inputs for a TI method}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15 rows and 2 columns.}
\usage{
allowed_inputs
}
\description{
All allowed inputs for a TI method
}
\examples{
allowed_inputs

}
\keyword{create_ti_method}
