% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allowed_outputs}
\alias{allowed_outputs}
\title{All allowed outputs}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 5 columns.}
\usage{
allowed_outputs
}
\description{
All allowed outputs
}
\keyword{datasets}
