% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_attraction.R
\name{calculate_attraction}
\alias{calculate_attraction}
\title{Calculate the attraction of cells to other cells using velocity}
\usage{
calculate_attraction(current, projected, cells = colnames(projected),
  n_waypoints = 50, k = 50, debug = FALSE)
}
\arguments{
\item{current}{Current expression}

\item{projected}{Projected expression based on RNA velocity}
}
\value{
Matrix containing the attraction (\link{-1, 1}) of each cell to the waypoint cells
}
\description{
Calculate the attraction of cells to other cells using velocity
}
