% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_geodesic_distances.R
\name{calculate_geodesic_distances}
\alias{calculate_geodesic_distances}
\alias{compute_tented_geodesic_distances}
\title{Calculate geodesic distances between cells in a trajectory}
\usage{
calculate_geodesic_distances(trajectory, waypoint_cells = NULL,
  waypoint_milestone_percentages = NULL)

compute_tented_geodesic_distances(trajectory, waypoint_cells = NULL,
  waypoint_milestone_percentages = NULL)
}
\arguments{
\item{trajectory}{The trajectory}

\item{waypoint_cells}{A vector of waypoint cells. Only the geodesic distances between waypoint cells and all other cells will be calculated.}

\item{waypoint_milestone_percentages}{The milestone percentages of non-cell waypoints, containing waypoint_id, milestone_id and percentage columns}
}
\description{
Will calculate geodesic distances along a trajectory. To speed things up, only the distances with a set of waypoint cells are calculated.
}
\details{
The geodesic distance takes into acount regions of delayed commitment.
}
\keyword{derive_trajectory}
