% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{common_param}
\alias{common_param}
\title{Common param}
\arguments{
\item{trajectory}{The trajectory as created by \code{\link[=infer_trajectory]{infer_trajectory()}} or \code{\link[=add_trajectory]{add_trajectory()}}}

\item{dataset}{A dataset created by \code{\link[=wrap_data]{wrap_data()}} or \code{\link[=wrap_expression]{wrap_expression()}}}
}
\description{
Common param
}
\keyword{internal}
