% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_tented_geodesic_distances.R
\name{compute_tented_geodesic_distances}
\alias{compute_tented_geodesic_distances}
\alias{compute_tented_geodesic_distances_}
\title{Calculate geodesic distances between cells in a trajectory, taking into account tents}
\usage{
compute_tented_geodesic_distances(trajectory, waypoint_cells = NULL,
  waypoint_milestone_percentages = NULL)

compute_tented_geodesic_distances_(cell_ids, milestone_ids,
  milestone_network, milestone_percentages, divergence_regions,
  waypoint_cells = NULL, waypoint_milestone_percentages = NULL)
}
\arguments{
\item{trajectory}{The trajectory}

\item{waypoint_cells}{A vector of waypoint cells. Only the geodesic distances between waypoint cells and all other cells will be calculated.}

\item{waypoint_milestone_percentages}{The milestone percentages of non-cell waypoints, containing waypoint_id, milestone_id and percentage columns}

\item{cell_ids}{The ids of the cells.}

\item{milestone_ids}{The ids of the milestones in the trajectory. Type: Character vector.}

\item{milestone_network}{The network of the milestones.
Type: Data frame(from = character, to = character, length = numeric, directed = logical).}

\item{milestone_percentages}{A data frame specifying what percentage milestone each cell
consists of.
Type: Data frame(cell_id = character, milestone_id = character, percentage = numeric).}

\item{divergence_regions}{A data frame specifying the divergence
regions between milestones (e.g. a bifurcation).
Type: Data frame(divergence_id = character, milestone_id = character, is_start = logical).}
}
\description{
Calculate geodesic distances between cells in a trajectory, taking into account tents
}
\keyword{derive_trajectory}
