% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_config.R
\name{container_config}
\alias{container_config}
\alias{container_docker}
\alias{container_singularity}
\title{Backend configuration for containerisation}
\usage{
container_config(type = getOption("dynwrap_run_environment") \%||\%
  getenv("DYNWRAP_RUN_ENVIRONMENT") \%||\% "docker", ...)

container_docker()


  container_singularity(images_folder = getOption("dynwrap_singularity_images_folder")
  \%||\% getenv("DYNWRAP_SINGULARITY_IMAGES_FOLDER") \%||\% "./")
}
\arguments{
\item{type}{Which backend to use. Can be either \code{"docker"} or \code{"singularity"}.
If the value of this parameter is not passed explicitly, the default depend on
the \code{"dynwrap_run_environment"} option, the \code{"DYNWRAP_RUN_ENVIRONMENT"} environment
variable, or otherwise just \code{"docker"}.}

\item{...}{Parameters to pass to \code{container_docker()} or \code{container_singularity()}.}

\item{images_folder}{A folder in which to store the singularity images. Each TI method will require about 1 to 2GB of space.}
}
\description{
Backend configuration for containerisation
}
