% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_config.R
\name{container_create_config}
\alias{container_create_config}
\alias{container_create_docker_config}
\alias{container_create_singularity_config}
\alias{container_get_default_config}
\alias{container_set_default_config}
\title{Backend configuration for containerisation}
\usage{
container_create_config(
  type =
    getOption("dynwrap_run_environment") \%||\%
    get_env_or_null("DYNWRAP_RUN_ENVIRONMENT") \%||\%
    "docker",
  ...
)

container_create_docker_config()

container_create_singularity_config(
  images_folder =
    getOption("dynwrap_singularity_images_folder") \%||\%
    get_env_or_null("DYNWRAP_SINGULARITY_IMAGES_FOLDER") \%||\%
    "./",
  prebuild = images_folder != "./"
)

container_get_default_config()

container_set_default_config(config, permanent = TRUE)
}
\arguments{
\item{type}{Which backend to use. Can be either \code{"docker"} or \code{"singularity"}.
If the value of this parameter is not passed explicitly, the default depend on
the \code{"dynwrap_run_environment"} option, the \code{"DYNWRAP_RUN_ENVIRONMENT"} environment
variable, or otherwise just \code{"docker"}.}

\item{...}{Parameters to pass to \code{container_create_docker_config()} or \code{container_create_singularity_config()}.}

\item{images_folder}{A folder in which to store the singularity images. Each TI method will require about 1GB of space.}

\item{prebuild}{If the singularity images are not prebuilt, they will need to be built every time a method is run.}

\item{config}{A config to save as default.}

\item{permanent}{Whether or not to save the config file permanently}
}
\description{
Backend configuration for containerisation
}
\examples{
\dontrun{
config <- container_create_docker_config()
container_set_default_config(config)

config <- container_create_singularity_config(
  images_folder = "~/dynwrap_singularity_images/"
)
container_set_default_config(config)
}

}
