% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_process_definition.R
\name{convert_definition}
\alias{convert_definition}
\title{Convert a definition loaded in from a yaml}
\usage{
convert_definition(definition_raw)
}
\arguments{
\item{definition_raw}{The raw definition loaded from the yaml}
}
\description{
Convert a definition loaded in from a yaml
}
\keyword{create_ti_method}
