% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_milestone_percentages_to_progressions.R
\name{convert_milestone_percentages_to_progressions}
\alias{convert_milestone_percentages_to_progressions}
\title{Conversion between milestone percentages and progressions}
\usage{
convert_milestone_percentages_to_progressions(
  cell_ids,
  milestone_ids,
  milestone_network,
  milestone_percentages
)
}
\arguments{
\item{cell_ids}{The identifiers of the cells.}

\item{milestone_ids}{The ids of the milestones in the trajectory. Type: Character vector.}

\item{milestone_network}{The network of the milestones.
Type: Data frame(from = character, to = character, length = numeric, directed = logical).}

\item{milestone_percentages}{A data frame specifying what percentage milestone each cell
consists of.
Type: Data frame(cell_id = character, milestone_id = character, percentage = numeric).}
}
\value{
A data frame with columns \code{cell_id}, \code{from}, \code{to}, \code{percentage}.
}
\description{
Conversion between milestone percentages and progressions
}
\examples{
progressions <- convert_milestone_percentages_to_progressions(
  cell_ids = example_trajectory$cell_ids,
  milestone_ids = example_trajectory$milestone_ids,
  milestone_network = example_trajectory$milestone_network,
  milestone_percentages = example_trajectory$milestone_percentages
)
head(progressions)

}
\seealso{
\code{\link[=add_trajectory]{add_trajectory()}}, \link{convert_progressions_to_milestone_percentages}
}
