% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_create_docker_method.R
\name{create_docker_ti_method}
\alias{create_docker_ti_method}
\alias{pull_docker_ti_method}
\title{Create a TI method from a docker(hub) image}
\usage{
create_docker_ti_method(image, name = NULL, input = NULL,
  input_optional = NULL, output = NULL, par_set = NULL,
  parameters = NULL, definition = NULL,
  docker_client = stevedore::docker_client())

pull_docker_ti_method(image, docker_client = stevedore::docker_client())
}
\arguments{
\item{image}{The name of the image. Required}

\item{name}{The name of the method. Optional, will be extracted from the image definition.yml file}

\item{input}{Names of input data, such as expression, counts or some \link{priors}}

\item{input_optional}{Names of optional input data, such as expression, counts, or some \link{priors}}

\item{output}{Output generated by the method}

\item{par_set}{The parameter set as created by \link[ParamHelpers:makeParamSet]{ParamHelpers::makeParamSet}}

\item{parameters}{A named list of parameters, containing the type of parameter and additional parameters such as lower and upper bounds, default value and/or possible values. Optional, will be extracted from the image definition.yml file}

\item{definition}{A list containing the name, image, input, output and parameters of a method. Usually used for when these values are loaded from a yaml or json file}

\item{docker_client}{Optional, a \code{\link[stevedore:docker_client]{stevedore::docker_client()}}}
}
\description{
\code{create_docker_ti_method} will use a local docker image, \code{pull_docker_ti_method} will pull the latest docker image from the docker hub
}
