% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_create_ti_method_definition.R
\name{create_ti_method_definition}
\alias{create_ti_method_definition}
\title{Create a TI method from a local method definition file}
\usage{
create_ti_method_definition(filename = NULL, definition = NULL,
  return_function = TRUE)
}
\arguments{
\item{filename}{The filename of the method definition.}

\item{definition}{The definition if already available as a list. Exactly one of \code{filename} and \code{definition} must be NULL.}
}
\description{
The local method definition file describes a method that is runnable on the local system.
}
