% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_process_definition.R
\name{def_container}
\alias{def_container}
\title{Define the container}
\usage{
def_container(docker, url = NULL)
}
\arguments{
\item{docker}{docker}

\item{url}{url}
}
\description{
Define the container
}
