% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_process_definition.R
\name{def_manuscript}
\alias{def_manuscript}
\title{Define the manuscript}
\usage{
def_manuscript(doi = NULL, google_scholar_cluster_id = NULL,
  preprint_date = NULL, publication_date = NULL)
}
\arguments{
\item{doi}{doi}

\item{google_scholar_cluster_id}{google_scholar_cluster_id}

\item{preprint_date}{preprint_data}

\item{publication_date}{publication_date}
}
\description{
Define the manuscript
}
