% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_process_definition.R
\name{def_method}
\alias{def_method}
\title{Define the method}
\usage{
def_method(id, name = id, source = "tool", platform = "R",
  url = NULL, license = NULL, authors = list())
}
\arguments{
\item{id}{id}

\item{name}{name}

\item{source}{source}

\item{platform}{platform}

\item{url}{url}

\item{license}{license}

\item{authors}{authors}
}
\description{
Define the method
}
