% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_process_definition.R
\name{def_package}
\alias{def_package}
\title{Meta information on the package in which the TI function resides}
\usage{
def_package(remote, name, function_name)
}
\arguments{
\item{remote}{The github repository handle}

\item{name}{The name of the package}

\item{function_name}{The name of the function}
}
\description{
Meta information on the package in which the TI function resides
}
\examples{
def_package(
  remote = "rcannood/SCORPIUS",
  name = "SCORPIUS",
  function_name = "ti_scorpius"
)
}
\keyword{create_ti_method}
