% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_process_definition.R
\name{def_wrapper}
\alias{def_wrapper}
\title{Define the wrapper}
\usage{
def_wrapper(input_required, input_optional = character(),
  type = "trajectory", topology_inference = NULL,
  trajectory_types = character())
}
\arguments{
\item{input_required}{The required inputs for this method. See \code{dynwrap::allowed_inputs()}.}

\item{input_optional}{Optional inputs for this method. See \code{dynwrap::allowed_inputs()}.}

\item{type}{type}

\item{topology_inference}{topology_inference}

\item{trajectory_types}{trajectory_types}
}
\description{
Define the wrapper
}
