% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_process_definition.R
\name{definition}
\alias{definition}
\title{Create a definition}
\usage{
definition(method, wrapper, manuscript = NULL, container = NULL,
  parameters = parameter_set())
}
\arguments{
\item{method}{See \code{\link[=def_method]{def_method()}}}

\item{wrapper}{See \code{\link[=def_wrapper]{def_wrapper()}}}

\item{manuscript}{See \code{\link[=def_manuscript]{def_manuscript()}}}

\item{container}{See \code{\link[=def_container]{def_container()}}}

\item{parameters}{See \code{\link[=def_parameters]{def_parameters()}}}
}
\description{
Create a definition
}
