% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred_trajectory.R
\name{dimred_trajectory}
\alias{dimred_trajectory}
\alias{check_or_perform_dimred}
\title{Perform dimensionality reduction on a trajectory and the respective samples in order to plot it}
\usage{
dimred_trajectory(trajectory, adjust_weights = TRUE)

check_or_perform_dimred(object)
}
\arguments{
\item{trajectory}{The trajectory}

\item{object}{An object generated by \code{wrap_data} or \code{dimred_trajectory}.}
}
\description{
Perform dimensionality reduction on a trajectory and the respective samples in order to plot it

Perform dimred trajectory if needed
}
