% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{dynwrap-package}
\alias{dynwrap}
\alias{dynwrap-package}
\title{Inferring and adapting single-cell trajectories}
\description{
\figure{logo.png}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dynverse/dynwrap}
  \item Report bugs at \url{https://github.com/dynverse/dynwrap/issues}
}

}
\author{
\strong{Maintainer}: Robrecht Cannoodt \email{rcannood@gmail.com} (\href{https://orcid.org/0000-0003-3641-729X}{ORCID})

Authors:
\itemize{
  \item Wouter Saelens \email{wouter.saelens@gmail.com} (\href{https://orcid.org/0000-0002-7114-6248}{ORCID})
}

}
\keyword{internal}
