% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dataset}
\alias{example_dataset}
\title{Example dataset}
\format{An object of class \code{dynwrap::with_dimred} (inherits from \code{dynwrap::with_expression}, \code{dynwrap::data_wrapper}, \code{list}) of length 12.}
\usage{
example_dataset
}
\description{
Example dataset
}
\keyword{datasets}
