% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_infer_trajectory.R
\name{execute_method_internal}
\alias{execute_method_internal}
\title{Internal method for executing a method}
\usage{
execute_method_internal(method, arglist, time0)
}
\arguments{
\item{method}{One or more methods. Must be one of:
\itemize{
\item{an object or list of ti_... objects (eg. \code{\link[=ti_comp1]{ti_comp1()}})}
\item{a character vector containing the names of methods to execute (e.g. \code{"scorpius"}), or}
\item{a dynguidelines data frame.}
}}

\item{arglist}{The arguments to apply to the method}

\item{time0}{The start of the timer.}
}
\description{
If you're reading this, you're supposed to be using \code{infer_trajectory} instead.
}
