% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_infer_trajectory.R
\name{execute_method_on_dataset}
\alias{execute_method_on_dataset}
\title{Helper function for running a method on a set of parameters}
\usage{
execute_method_on_dataset(dataset, method, parameters = NULL,
  give_priors = NULL, verbose = FALSE, capture_output, seed = 1)
}
\arguments{
\item{dataset}{The dataset}

\item{method}{One or more methods. Must be one of:
\itemize{
\item{an object or list of ti_... objects (eg. \code{dynmethods::ti_comp1()}),}
\item{a character vector containing the names of methods to execute (e.g. \code{"scorpius"}),}
\item{a character vector containing dockerhub repositories (e.g. \code{dynverse/paga}), or}
\item{a dynguidelines data frame.}
}}

\item{parameters}{A set of parameters to be used during trajectory inference.
A parameter set must be a named list of parameters.
If multiple methods were provided in the \code{method} parameter,
\code{parameters} must be an unnamed list of the same length.}

\item{give_priors}{All the priors a method is allowed to receive.
Must be a subset of all available priors (\code{\link[dynwrap:priors]{priors}}).}

\item{verbose}{Whether or not to print information output.}

\item{capture_output}{Whether to capture the stdout and stderr produced by a method.}

\item{seed}{A seed to be set, if the method allows for it.}
}
\description{
Helper function for running a method on a set of parameters
}
