% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_orient_topology.R
\name{flip_edges}
\alias{flip_edges}
\title{Flip a set of edges of the milestone network}
\usage{
flip_edges(trajectory, milestone_network_toflip)
}
\arguments{
\item{trajectory}{The trajectory as created by \code{\link[=infer_trajectory]{infer_trajectory()}} or \code{\link[=add_trajectory]{add_trajectory()}}}

\item{milestone_network_toflip}{A dataframe with a from and to column, containing the subset of the milestone network #'}
}
\description{
Note that this will remove associated roots, reroot the trajectory using \code{\link[=add_root]{add_root()}}
}
\keyword{adapt_trajectory}
