% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_gather_cells_at_milestones.R
\name{gather_cells_at_milestones}
\alias{gather_cells_at_milestones}
\title{"Gather" cells to their closest milestones}
\usage{
gather_cells_at_milestones(trajectory)
}
\arguments{
\item{trajectory}{The trajectory}
}
\description{
"Gather" cells to their closest milestones
}
