% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_create_ti_method_r.R
\name{generate_parameter_documentation}
\alias{generate_parameter_documentation}
\title{Generate the parameter documentation of a method, use with \code{@eval}}
\usage{
generate_parameter_documentation(definition)
}
\arguments{
\item{definition}{The definition which contain the parameters}
}
\value{
A character vector containing the roxygen tags
}
\description{
Generate the parameter documentation of a method, use with \code{@eval}
}
