% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_grouping.R
\name{get_cell_grouping}
\alias{get_cell_grouping}
\title{Group cells to their highest milestone}
\usage{
get_cell_grouping(milestone_percentages)
}
\arguments{
\item{milestone_percentages}{milestone percentages dataframe}
}
\description{
Group cells to their highest milestone
}
