% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ti_method.R
\name{get_default_parameters}
\alias{get_default_parameters}
\title{Get the default parameters of a method}
\usage{
get_default_parameters(method)
}
\arguments{
\item{method}{A TI method description}
}
\description{
Get the default parameters of a method
}
