% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_divergence_triangles.R
\name{get_divergence_triangles}
\alias{get_divergence_triangles}
\title{Helper function for processing divergence regions}
\usage{
get_divergence_triangles(divergence_regions)
}
\arguments{
\item{divergence_regions}{A divergence regions data frame as produced by \code{add_trajectory}.}
}
\description{
This function returns the combinations between
the start of each divergence region and pairwise combinations
of the end milestones.
}
