% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_expression.R
\name{get_expression}
\alias{get_expression}
\title{Get expression}
\usage{
get_expression(task, expression_source)
}
\arguments{
\item{task}{The task}

\item{expression_source}{`expression`, `counts` or a matrix}
}
\description{
Get expression
}
