% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_grouping.R
\name{group_counts}
\alias{group_counts}
\title{Calculate mean expression levels for every cell group}
\usage{
group_counts(counts, cell_grouping)
}
\arguments{
\item{counts}{matrix}

\item{cell_grouping}{grouping of cells dataframe}
}
\description{
Calculate mean expression levels for every cell group
}
