% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_grouping.R
\name{group_from_trajectory}
\alias{group_from_trajectory}
\alias{group_onto_trajectory_edges}
\alias{group_onto_nearest_milestones}
\title{Create a grouping from a trajectory}
\usage{
group_onto_trajectory_edges(trajectory, group_template = "{from}->{to}")

group_onto_nearest_milestones(trajectory)
}
\arguments{
\item{trajectory}{The trajectory as created by \code{\link[=infer_trajectory]{infer_trajectory()}} or \code{\link[=add_trajectory]{add_trajectory()}}}

\item{group_template}{Processed by glue::glue to name the group}
}
\description{
Grouping cells onto their edges, or grouping cells onto their nearest milestones
}
\keyword{derive_trajectory}
