% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_grouping.R
\name{group_onto_nearest_milestones}
\alias{group_onto_nearest_milestones}
\title{Grouping the cells onto the closest milestones}
\usage{
group_onto_nearest_milestones(trajectory)
}
\arguments{
\item{trajectory}{The trajectory object}
}
\description{
Grouping the cells onto the closest milestones
}
