% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_grouping.R
\name{group_onto_trajectory_edges}
\alias{group_onto_trajectory_edges}
\title{Grouping the cells onto their edges}
\usage{
group_onto_trajectory_edges(trajectory, group_template = "{from}->{to}")
}
\arguments{
\item{trajectory}{The trajectory object}

\item{group_template}{Processed by glue::glue to name the group}
}
\description{
Grouping the cells onto their edges
}
