% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_infer_trajectory.R
\name{infer_trajectories}
\alias{infer_trajectories}
\alias{infer_trajectory}
\title{Infer trajectories}
\usage{
infer_trajectories(dataset, method, parameters = NULL,
  give_priors = NULL, seed = random_seed(), verbose = FALSE,
  return_verbose = FALSE, debug = FALSE, map_fun = map)

infer_trajectory(dataset, method, parameters = NULL,
  give_priors = NULL, seed = random_seed(), verbose = FALSE,
  return_verbose = FALSE, debug = FALSE, ...)
}
\arguments{
\item{dataset}{One or more datasets, as created using dynwrap.}

\item{method}{One or more methods. Must be one of:
\itemize{
\item{an object or list of ti_... objects (eg. \code{dynmethods::ti_comp1()}),}
\item{a character vector containing the names of methods to execute (e.g. \code{"scorpius"}),}
\item{a character vector containing dockerhub repositories (e.g. \code{dynverse/paga}), or}
\item{a dynguidelines data frame.}
}}

\item{parameters}{A set of parameters to be used during trajectory inference.
A parameter set must be a named list of parameters.
If multiple methods were provided in the \code{method} parameter,
\code{parameters} must be an unnamed list of the same length.}

\item{give_priors}{All the priors a method is allowed to receive.
Must be a subset of all available priors (\code{\link[dynwrap:priors]{priors}}).}

\item{seed}{A seed to be passed to the TI method.}

\item{verbose}{Whether or not to print information output.}

\item{return_verbose}{Whether to store and return messages printed by the method.}

\item{debug}{Used for debugging containers methods.}

\item{map_fun}{A mao function to use when inferring trajectories with multiple datasets or methods.
Allows to parallellise the execution in an arbitrary way.}

\item{...}{Any additional parameters given to the method, will be concatednated to the parameters argument}
}
\description{
Infer trajectories
}
\keyword{infer_trajectory}
