% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_cell_group_to_wrapper.R
\name{is_wrapper_with_cell_group}
\alias{is_wrapper_with_cell_group}
\title{Test whether an object is a data_wrapper and has cell_group data}
\usage{
is_wrapper_with_cell_group(object)
}
\arguments{
\item{object}{The object to be tested.}
}
\description{
Test whether an object is a data_wrapper and has cell_group data
}
