% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_waypoints_to_wrapper.R
\name{is_wrapper_with_waypoints}
\alias{is_wrapper_with_waypoints}
\title{Test whether an object is a data_wrapper and waypoints}
\usage{
is_wrapper_with_waypoints(object)
}
\arguments{
\item{object}{The object to be tested.}
}
\description{
Test whether an object is a data_wrapper and waypoints
}
