% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_label_milestones.R
\name{label_milestones}
\alias{label_milestones}
\alias{label_milestones_markers}
\alias{is_wrapper_with_milestone_labelling}
\alias{get_milestone_labelling}
\title{Label milestones either manually (\code{label_milestones}) or using marker genes (\code{label_milestones_markers})}
\usage{
label_milestones(trajectory, labelling)

label_milestones_markers(trajectory, markers,
  expression_source = "expression", n_nearest_cells = 20)

is_wrapper_with_milestone_labelling(trajectory)

get_milestone_labelling(trajectory, label_milestones = NULL)
}
\arguments{
\item{trajectory}{The trajectory}

\item{labelling}{Named character vector containing for a milestone a new label}

\item{markers}{List containing for each label a list of marker genes}

\item{expression_source}{The expression source}

\item{n_nearest_cells}{The number of nearest cells to use for extracting milestone expression}

\item{label_milestones}{How to label the milestones. Can be TRUE (in which case the labels within the trajectory will be used), "all" (in which case both given labels and milestone_ids will be used), a named character vector, or FALSE}
}
\description{
Label milestones either manually (\code{label_milestones}) or using marker genes (\code{label_milestones_markers})
}
\keyword{adapt_trajectory}
