% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_process_definition.R
\name{read_definition}
\alias{read_definition}
\title{Read a definition from a yaml file}
\usage{
read_definition(file)
}
\arguments{
\item{file}{The file name}
}
\description{
Read a definition from a yaml file
}
