% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{test_docker_installation}
\alias{test_singularity_installation}
\alias{create_config}
\alias{create_docker_config}
\alias{create_singularity_config}
\alias{get_default_config}
\alias{set_default_config}
\alias{test_h5_installation}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{babelwhale}{\code{\link[babelwhale]{create_config}}, \code{\link[babelwhale:create_config]{create_docker_config}}, \code{\link[babelwhale:create_config]{create_singularity_config}}, \code{\link[babelwhale:create_config]{get_default_config}}, \code{\link[babelwhale:create_config]{set_default_config}}, \code{\link[babelwhale]{test_docker_installation}}, \code{\link[babelwhale]{test_singularity_installation}}}

  \item{dynutils}{\code{\link[dynutils]{test_h5_installation}}}
}}

