% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_waypoints_to_wrapper.R
\name{select_waypoints}
\alias{select_waypoints}
\title{Select the waypoints}
\usage{
select_waypoints(traj, resolution = 0.1)
}
\arguments{
\item{traj}{The trajectory object}

\item{resolution}{The resolution of the waypoints, measured in the same units as the lengths of the milestone network edges}
}
\description{
Waypoints are spread equally over the whole trajectory
}
