% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_network.R
\name{simplify_milestone_network}
\alias{simplify_milestone_network}
\title{Simplify a milestone network, so that consecutive linear edges are merged}
\usage{
simplify_milestone_network(net)
}
\arguments{
\item{net}{A network in data.frame format containing at least the columns from, to and length}
}
\description{
Simplify a milestone network, so that consecutive linear edges are merged
}
\examples{
net <- data.frame(from=1:2, to=2:3, length=1, directed=TRUE, stringsAsFactors = F)
simplify_milestone_network(net)

net <- data.frame(from=c(1,2,3, 4), to=c(2,3,4, 1), directed=TRUE, length=1)
simplify_milestone_network(net)
}
