% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_trajectory.R
\name{simplify_trajectory}
\alias{simplify_trajectory}
\title{Simplify a trajectory by removing transient milestones}
\usage{
simplify_trajectory(trajectory, allow_self_loops = FALSE)
}
\arguments{
\item{trajectory}{The trajectory as created by \code{\link[=infer_trajectory]{infer_trajectory()}} or \code{\link[=add_trajectory]{add_trajectory()}}}

\item{allow_self_loops}{Whether or not to allow self loops.}
}
\description{
Transient milestones are milestones with a degree of two (if undirected) or with an indegree of one and outdegree of one (if directed)
}
\details{
The positions of the cells within the trajectory remain the same

A → B → C is simplified to A → C.

A — B — C is simplified to A — C.

A → B → C → D → A is simplified to A → B → D.
}
\keyword{adapt_trajectory}
