% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_milestone_network.R
\name{simplify_trajectory_type}
\alias{simplify_trajectory_type}
\title{Convert directed trajectory type to simplified versions}
\usage{
simplify_trajectory_type(trajectory_types)
}
\arguments{
\item{trajectory_types}{trajectory types to simplify}
}
\description{
Convert directed trajectory type to simplified versions
}
