% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_gene_importances.R
\name{trajectory_gene_importances}
\alias{trajectory_gene_importances}
\title{Attempt at calculating gene importances per milestone}
\usage{
trajectory_gene_importances(traj, mtry = function(x) ncol(x) * 0.01,
  num_trees = 10000, num_cores = 1)
}
\arguments{
\item{traj}{A trajectory object containing expression values and a trajectory.}

\item{mtry}{A function for determining the mtry parameter.}

\item{num_trees}{The number of trees to use.}

\item{num_cores}{The number of cores to use.}
}
\description{
Uses the feature importance measures of \code{\link[ranger]{ranger}}.
}
