% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_output_processors.R
\name{wrap_output}
\alias{wrap_output}
\title{Wrap the output of a TI method}
\usage{
wrap_output(model, output_ids, dir_output)
}
\arguments{
\item{model}{The model to start from, as generated by \code{wrap_data}}

\item{output_ids}{The ids of the outputs generated by the methods}

\item{dir_output}{The directory containing the output files}
}
\description{
Wrap the output of a TI method
}
