% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_output_list.R
\name{wrap_output_list}
\alias{wrap_output_list}
\title{Transform a list of data objects to a dynwrap trajectory}
\usage{
wrap_output_list(output, output_ids)
}
\arguments{
\item{output}{A list containing dynwrap data}

\item{output_ids}{The names of the promised dynwrap objects}
}
\description{
Transform a list of data objects to a dynwrap trajectory
}
