% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wrapper_types}
\alias{wrapper_types}
\title{Metadata on wrapper types}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 3 columns.}
\usage{
wrapper_types
}
\description{
Metadata on wrapper types
}
\keyword{datasets}
