% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_branch_trajectory.R
\name{add_branch_trajectory}
\alias{add_branch_trajectory}
\title{Create a trajectory given its branch network and the pseudotime of the cells on one of the branches}
\usage{
add_branch_trajectory(dataset, branch_network, branches,
  branch_progressions, ...)
}
\arguments{
\item{dataset}{A dataset created by \code{\link[=wrap_data]{wrap_data()}} or \code{\link[=wrap_expression]{wrap_expression()}}}

\item{branch_network}{The network between branches
Type: Data frame(from = charactor, to = character)}

\item{branches}{The length and directedness of the branches
Type: Data frame(branch_id = character, length = numeric, directed = logical)}

\item{branch_progressions}{Specifies the progression of a cell along a transition in the branch network.
Type: Data frame(cell_id = character, branch_id = character, percentage = numeric).}

\item{...}{extra information to be stored in the trajectory}
}
\value{
The trajectory
}
\description{
Create a trajectory given its branch network and the pseudotime of the cells on one of the branches
}
\examples{
library(tibble)
dataset <- wrap_data(cell_ids = letters)

branch_network <- tibble(from = c("A", "A"), to = c("B", "C"))
branch_network
branches <- tibble(branch_id = c("A", "B", "C"), length = 1, directed = TRUE)
branches
branch_progressions <- tibble(
  cell_id = dataset$cell_ids,
  branch_id = sample(branches$branch_id, length(dataset$cell_ids), replace = TRUE),
  percentage = runif(length(dataset$cell_ids))
)
branch_progressions

trajectory <- add_branch_trajectory(
  dataset,
  branch_network,
  branches,
  branch_progressions
)

if ("dynplot" \%in\% rownames(installed.packages())) {dynplot::plot_graph(trajectory)}
}
\keyword{create_trajectory}
