% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_cell_graph.R
\name{add_cell_graph}
\alias{add_cell_graph}
\title{Constructs a trajectory using a graph between cells, by mapping cells onto a set of backbone cells.}
\usage{
add_cell_graph(dataset, cell_graph, to_keep,
  milestone_prefix = "milestone_", ...)
}
\arguments{
\item{dataset}{A dataset created by \code{\link[=wrap_data]{wrap_data()}} or \code{\link[=wrap_expression]{wrap_expression()}}}

\item{cell_graph}{The edges between cells. Format: Data frame(from = character, to = character, length(optional) = numeric, directed(optional) = logical)}

\item{to_keep}{A named vector containing booleans containing
whether or not a cell is part of the backbone. Or, alternatively a character vector containing the backbone cells}

\item{milestone_prefix}{A prefix to add to the id of the cell ids when they are used as milestones, in order to avoid any naming conflicts,}

\item{...}{extra information to be stored in the wrapper.}
}
\value{
The trajectory
}
\description{
This function will generate the milestone_network and progressions.
}
\keyword{create_trajectory}
