% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_cluster_graph.R
\name{add_cluster_graph}
\alias{add_cluster_graph}
\title{Add a cluster projection trajectory to a data wrapper}
\usage{
add_cluster_graph(model, milestone_network, ...)
}
\arguments{
\item{model}{The model to which a cluster graph will be added. Needs to have a cell grouping created by \code{\link[=add_grouping]{add_grouping()}}.}

\item{milestone_network}{A network of milestones.}

\item{...}{extra information to be stored in the wrapper.}
}
\value{
The trajectory model
}
\description{
This function will generate the milestone_network and progressions.
}
