% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_dimred.R
\name{add_dimred}
\alias{add_dimred}
\alias{is_wrapper_with_dimred}
\alias{get_dimred}
\title{Add or create a dimensionality reduction}
\usage{
add_dimred(model, dimred, dimred_milestones = NULL,
  dimred_segment_progressions = NULL, dimred_segment_points = NULL,
  connect_segments = FALSE, expression_source = "expression", ...)

is_wrapper_with_dimred(model)

get_dimred(model, dimred = NULL, expression_source = "expression")
}
\arguments{
\item{model}{The model to which a dimensionality reduction will be added.}

\item{dimred}{The dimensionality reduction matrix (with cell_ids as rownames) or function which will run the dimensionality reduction}

\item{dimred_milestones}{An optional dimensionality reduction of the milestones.}

\item{dimred_segment_progressions}{An optional progression matrix of the trajectory segments. Format: \code{tibble(from, to, percentage)}}

\item{dimred_segment_points}{An optional dimensionality reduction of the trajectory segments. Format: \code{matrix(Comp1, Comp2, ...)}.}

\item{connect_segments}{Whether to connect segments between edges}

\item{expression_source}{The source of expression, can be "counts", "expression", an expression matrix, or another model which contains expression}

\item{...}{extra information to be stored in the wrapper}
}
\description{
Add or create a dimensionality reduction
}
\keyword{adapt_trajectory}
