% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_dimred.R
\name{add_dimred}
\alias{add_dimred}
\alias{is_wrapper_with_dimred}
\alias{get_dimred}
\title{Add or create a dimensionality reduction}
\usage{
add_dimred(dataset, dimred, dimred_milestones = NULL,
  dimred_segment_progressions = NULL, dimred_segment_points = NULL,
  connect_segments = FALSE, project_trajectory = TRUE,
  expression_source = "expression", ...)

is_wrapper_with_dimred(dataset)

get_dimred(dataset, dimred = NULL, expression_source = "expression")
}
\arguments{
\item{dataset}{A dataset created by \code{\link[=wrap_data]{wrap_data()}} or \code{\link[=wrap_expression]{wrap_expression()}}}

\item{dimred}{The dimensionality reduction matrix (with cell_ids as rownames) or function which will run the dimensionality reduction}

\item{dimred_milestones}{An optional dimensionality reduction of the milestones.}

\item{dimred_segment_progressions}{An optional progression matrix of the trajectory segments. Format: \code{tibble(from, to, percentage)}}

\item{dimred_segment_points}{An optional dimensionality reduction of the trajectory segments. Format: \code{matrix(comp_1, comp_2, ...)}.}

\item{connect_segments}{Whether to connect segments between edges}

\item{project_trajectory}{Whether to also project the trajectory. Only relevant if dataset contains a trajectory, and dimred_segment_progressions and dimred_segment_points are not provided}

\item{expression_source}{The source of expression, can be "counts", "expression", an expression matrix, or another dataset which contains expression}

\item{...}{extra information to be stored in the wrapper}
}
\description{
Add or create a dimensionality reduction
}
\keyword{adapt_trajectory}
