% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_dimred_projection.R
\name{add_dimred_projection}
\alias{add_dimred_projection}
\title{Constructs a trajectory by projecting cells within a dimensionality reduction onto a backbone formed by a milestone network. Optionally, a cell grouping can be given which will restrict the edges on which a cell can be projected.}
\usage{
add_dimred_projection(dataset, milestone_ids = NULL, milestone_network,
  dimred, dimred_milestones, grouping = NULL, ...)
}
\arguments{
\item{dataset}{A dataset created by \code{\link[=wrap_data]{wrap_data()}} or \code{\link[=wrap_expression]{wrap_expression()}}}

\item{milestone_ids}{The ids of the milestones in the trajectory. Type: Character vector.}

\item{milestone_network}{The network of the milestones.
Type: Data frame(from = character, to = character, length = numeric, directed = logical).}

\item{dimred}{The dimensionality reduction matrix (with cell_ids as rownames) or function which will run the dimensionality reduction}

\item{dimred_milestones}{An optional dimensionality reduction of the milestones.}

\item{grouping}{A grouping of the cells, can be a named vector or a dataframe with group_id (character) and cell_id (character)}

\item{...}{extra information to be stored in the wrapper.}
}
\value{
The trajectory
}
\description{
This function will generate the milestone_network and progressions.
}
\keyword{create_trajectory}
