% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_expression.R
\name{add_expression}
\alias{add_expression}
\alias{is_wrapper_with_expression}
\alias{get_expression}
\title{Add count and normalised expression values to a model}
\usage{
add_expression(model, counts, expression, feature_info = NULL, ...)

is_wrapper_with_expression(model)

get_expression(model, expression_source = "expression")
}
\arguments{
\item{model}{The model to which expression will be added}

\item{counts}{The counts with genes in columns and cells in rows}

\item{expression}{The normalised expression values with genes in columns and cells in rows}

\item{feature_info}{Optional meta-information of the features, a data.frame with at least feature_id as column}

\item{...}{extra information to be stored in the model}

\item{expression_source}{The source of expression, can be "counts", "expression", an expression matrix, or another model which contains expression}
}
\description{
Add count and normalised expression values to a model
}
\keyword{adapt_trajectory}
