% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_grouping.R
\name{add_grouping}
\alias{add_grouping}
\alias{is_wrapper_with_grouping}
\alias{get_grouping}
\title{Add a cell grouping to a data wrapper}
\usage{
add_grouping(model, grouping, group_ids = NULL, ...)

is_wrapper_with_grouping(model)

get_grouping(model, grouping = NULL)
}
\arguments{
\item{model}{The model to which the grouping will be added.}

\item{grouping}{A grouping of the cells, can be a named vector or a dataframe with group_id (character) and cell_id (character)}

\item{group_ids}{All group_ids, optional}

\item{...}{Extra information to be stored in the wrapper.}
}
\description{
Add a cell grouping to a data wrapper
}
\keyword{adapt_trajectory}
