% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_pseudotime.R
\name{calculate_pseudotime}
\alias{calculate_pseudotime}
\alias{add_pseudotime}
\title{Add or calculate pseudotime as distance from the root}
\usage{
calculate_pseudotime(trajectory)

add_pseudotime(trajectory, pseudotime = NULL)
}
\arguments{
\item{trajectory}{The trajectory}

\item{pseudotime}{Named vector containing the pseudotime for every cell}
}
\description{
Add or calculate pseudotime as distance from the root
}
\keyword{derive_trajectory}
