% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_pseudotime.R
\name{calculate_pseudotime}
\alias{calculate_pseudotime}
\alias{add_pseudotime}
\title{Add or calculate pseudotime as distance from the root}
\usage{
calculate_pseudotime(trajectory)

add_pseudotime(trajectory, pseudotime = NULL)
}
\arguments{
\item{trajectory}{The trajectory as created by \code{\link[=infer_trajectory]{infer_trajectory()}} or \code{\link[=add_trajectory]{add_trajectory()}}}

\item{pseudotime}{Named vector containing the pseudotime for every cell. If not given, the pseudotime will be calculated.}
}
\value{
The trajectory with \emph{pseudotime} added, which is a named vector containing the pseudotime values for every cell.
}
\description{
When calculating the pseudotime, the trajectory is expected to be rooted (see \code{\link[=add_root]{add_root()}})
}
\seealso{
\code{\link[=add_root]{add_root()}}, \code{\link[=add_linear_trajectory]{add_linear_trajectory()}}
}
\keyword{derive_trajectory}
