% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root_trajectory.R
\name{add_root_using_expression}
\alias{add_root_using_expression}
\title{Add root cell to wrapper using expression of features}
\usage{
add_root_using_expression(task, features_oi, expression_source = "expression")
}
\arguments{
\item{task}{The task}

\item{feature_ids}{The feature ids}
}
\description{
Add root cell to wrapper using expression of features
}
