% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_timings.R
\name{add_timings}
\alias{add_timings}
\alias{is_wrapper_with_timings}
\alias{add_timing_checkpoint}
\title{Add count and normalised expression values to a model}
\usage{
add_timings(model, timings)

is_wrapper_with_timings(model)

add_timing_checkpoint(timings, name)
}
\arguments{
\item{model}{The model to which timings will be added}

\item{timings}{A list of timings.}

\item{name}{The name of the timings checkpoint.}
}
\description{
Add count and normalised expression values to a model

Helper function for storing timings information.
}
\keyword{adapt_trajectory}
