% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_timings.R
\name{add_timings}
\alias{add_timings}
\alias{is_wrapper_with_timings}
\alias{add_timing_checkpoint}
\title{Add timings to a trajectory}
\usage{
add_timings(trajectory, timings)

is_wrapper_with_timings(trajectory)

add_timing_checkpoint(timings, name)
}
\arguments{
\item{trajectory}{The trajectory as created by \code{\link[=infer_trajectory]{infer_trajectory()}} or \code{\link[=add_trajectory]{add_trajectory()}}}

\item{timings}{A list of timings.}

\item{name}{The name of the timings checkpoint.}
}
\description{
Add timings to a trajectory

Helper function for storing timings information.
}
\examples{
trajectory <- example_trajectory
trajectory <- add_timings(
  trajectory,
  list(start = 0, end = 1)
)

}
\keyword{adapt_trajectory}
