% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_trajectory.R
\name{add_trajectory}
\alias{add_trajectory}
\alias{write_milestone_network}
\alias{write_milestone_ids}
\alias{write_milestone_percentages}
\alias{write_progressions}
\alias{write_divergence_regions}
\title{Add a trajectory to a data wrapper}
\usage{
add_trajectory(data_wrapper, milestone_ids, milestone_network,
  divergence_regions, milestone_percentages = NULL, progressions = NULL,
  ...)

write_milestone_network(milestone_network, dir_output)

write_milestone_ids(milestone_ids, dir_output)

write_milestone_percentages(milestone_percentages, dir_output)

write_progressions(progressions, dir_output)

write_divergence_regions(divergence_regions, dir_output)
}
\arguments{
\item{data_wrapper}{A data wrapper to extend upon.}

\item{milestone_ids}{The ids of the milestones in the trajectory. Type: Character vector.}

\item{milestone_network}{The network of the milestones.
Type: Data frame(from = character, to = character, length = numeric, directed = logical).}

\item{divergence_regions}{A data frame specifying the divergence
regions between milestones (e.g. a bifurcation).
Type: Data frame(divergence_id = character, milestone_id = character, is_start = logical).}

\item{milestone_percentages}{A data frame specifying what percentage milestone each cell
consists of.
Type: Data frame(cell_id = character, milestone_id = character, percentage = numeric).}

\item{progressions}{Specifies the progression of a cell along a transition in the milestone_network.
Type: Data frame(cell_id = character, from = character, to = character, percentage = numeric).}

\item{...}{extra information to be stored in the wrapper.}

\item{dir_output}{The output directory}
}
\description{
Add a trajectory to a data wrapper
}
