% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_waypoints_to_wrapper.R
\name{add_waypoints_to_wrapper}
\alias{add_waypoints_to_wrapper}
\title{Add waypoints to a wrapped object with trajectory}
\usage{
add_waypoints_to_wrapper(object, resolution = 0.1)
}
\arguments{
\item{object}{Wrapper with trajectory}
}
\description{
Add waypoints to a wrapped object with trajectory
}
