% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_trajectory_dimred.R
\name{calculate_trajectory_dimred}
\alias{calculate_trajectory_dimred}
\title{Calculate a dimensionality reduction on a trajectory and its cells}
\usage{
calculate_trajectory_dimred(trajectory, adjust_weights = TRUE)
}
\arguments{
\item{trajectory}{A dynwrap trajectory}

\item{adjust_weights}{Whether or not to rescale the milestone network weights}
}
\description{
Calculate a dimensionality reduction on a trajectory and its cells
}
\seealso{
wrap_data
}
