% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_tented_geodesic_distances.R
\name{compute_tented_geodesic_distances}
\alias{compute_tented_geodesic_distances}
\title{Calculate geodesic distances between cells in a trajectory, taking into account tents}
\usage{
compute_tented_geodesic_distances(trajectory, waypoint_cells = NULL)
}
\arguments{
\item{trajectory}{The trajectory}

\item{waypoint_cells}{A vector of waypoint cells. Only the geodesic distances between waypoint cells and all other cells will be calculated.}
}
\description{
Calculate geodesic distances between cells in a trajectory, taking into account tents
}
