% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_create_docker_method.R
\name{create_docker_ti_method}
\alias{create_docker_ti_method}
\alias{pull_docker_ti_method}
\title{Create a TI method from a docker(hub) image}
\usage{
create_docker_ti_method(image, definition = load_definition_from_image(image,
  docker_client = docker_client), docker_client = stevedore::docker_client())

pull_docker_ti_method(image, docker_client = stevedore::docker_client())
}
\arguments{
\item{image}{The name of the image. Required}

\item{definition}{The method definition, a list containing the name, image, input, output and parameters of a method.
Optional, as the definition file will be automatically loaded from the image's \code{/code/definition.yml} using \code{\link{load_definition_from_image}}.}

\item{docker_client}{Optional, a \code{\link[stevedore:docker_client]{stevedore::docker_client()}}}
}
\description{
\code{create_docker_ti_method} will use a local docker image, \code{pull_docker_ti_method} will pull the latest docker image from the docker hub
}
