% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_create_method.R
\name{create_docker_ti_method}
\alias{create_docker_ti_method}
\alias{extract_definition_from_docker_image}
\alias{pull_docker_ti_method}
\title{Create a TI method from a docker image}
\usage{
create_docker_ti_method(image,
  definition = extract_definition_from_docker_image(image), ...)

extract_definition_from_docker_image(image,
  definition_location = "/code/definition.yml")

pull_docker_ti_method(image)
}
\arguments{
\item{image}{The name of the docker image, eg. \code{dynverse/comp1}. Can contain tags such as \code{dynverse/comp1:R_feather}}

\item{definition}{The method definition, a list containing the name, input, output and parameters of a method.
Optional, as the definition file will be automatically loaded from the images \code{/code/definition.yml} using \link{extract_definition_from_docker_image}.}

\item{...}{Other information about the method}

\item{definition_location}{The location of the definition file within the image}
}
\description{
These functions create a TI method from a docker image, either locally (\code{create_docker_ti_method}) or from \href{https://hub.docker.com/}{docker hub}  \code{pull_docker_ti_method}.
}
