% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_create_image_method.R
\name{create_singularity_ti_method}
\alias{create_singularity_ti_method}
\alias{extract_definition_from_singularity_image}
\title{Create a TI method from a singularity image}
\usage{
create_singularity_ti_method(image,
  definition = extract_definition_from_singularity_image(image), ...)

extract_definition_from_singularity_image(image,
  definition_location = "/code/definition.yml")
}
\arguments{
\item{image}{The location of the singularity image file, eg. \code{comp1.simg}.}

\item{definition}{The method definition, a list containing the name, input, output and parameters of a method.
Optional, as the definition file will be automatically loaded from the images \code{/code/definition.yml} using \link{extract_definition_from_singularity_image}.}

\item{...}{Other information about the method}

\item{definition_location}{The location of the definition file within the image}
}
\description{
This function creates a TI method from a singularity image. This image can be build using \code{singularity build} from a docker image (eg. \code{singularity build comp1.simg docker://dynverse/comp1})
}
