% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_create_ti_method_definition.R
\name{create_ti_method_definition}
\alias{create_ti_method_definition}
\title{Create a TI method from a local method definition file}
\usage{
create_ti_method_definition(definition, script, return_function = TRUE)
}
\arguments{
\item{definition}{A definition, see \code{\link[=definition]{definition()}}}

\item{script}{Location of the script that will be executed. Has to contain a #!}

\item{return_function}{Whether to return a function that allows you to override the default parameters, or just return the method meta data as is.}
}
\description{
The local method definition file describes a method that is runnable on the local system.
}
\keyword{create_ti_method}
