% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_create_ti_method_r.R
\name{create_ti_method_r}
\alias{create_ti_method_r}
\title{Create a TI method wrapper}
\usage{
create_ti_method_r(definition, run_fun, package_required = character(),
  package_loaded = character(), remotes_package = character(),
  return_function = TRUE)
}
\arguments{
\item{definition}{A definition, see \code{\link[=definition]{definition()}}}

\item{run_fun}{A function to infer a trajectory, with parameters counts/expression, params, priors, verbose and seed}

\item{package_required}{The packages that need to be installed before executing the method.}

\item{package_loaded}{The packages that need to be loaded before executing the method.}

\item{remotes_package}{Package from which the remote locations of dependencies have to be extracted, eg. \code{dynmethods}.}

\item{return_function}{Whether to return a function that allows you to override the default parameters, or just return the method meta data as is.}
}
\description{
Create a TI method wrapper
}
