% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_create_ti_method_with_container.R
\name{create_ti_method_with_container}
\alias{create_ti_method_with_container}
\title{Create a TI method from a docker image}
\usage{
create_ti_method_with_container(image, version = NULL,
  config = container_config(), pull_if_needed = TRUE)
}
\arguments{
\item{image}{The name of the docker repository (e.g. \code{"dynverse/angle"}).}

\item{version}{The minimum required version of the TI method container.
If the required version is higher than the currently installed version,
the container will be pulled from dockerhub or singularityhub.}

\item{config}{A container config. See \code{\link[=container_config]{container_config()}} for more information.}

\item{pull_if_needed}{Pull the image if not yet available.}
}
\description{
These functions create a TI method from a docker image. Supports both docker and singularity as a backend.
}
