% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_create_ti_method_with_container.R
\name{create_ti_method_with_container}
\alias{create_ti_method_with_container}
\title{Create a TI method from a docker image}
\usage{
create_ti_method_with_container(image, repo_digest,
  image_type = getOption("dynwrap_run_environment"),
  singularity_images_folder = get_default_singularity_images_folder())
}
\arguments{
\item{image}{The naame of the docker repository (e.g. \code{"dynverse/angle"}).}

\item{repo_digest}{Which hash to use (e.g. \code{"dynverse/angle@sha256:473e54..."}).
If \code{NULL}, the docker installation cannot be automatically updated.}

\item{image_type}{In which environment to run the method, can be \code{"docker"} or \code{"singularity"}.
By default, docker will be used.}

\item{singularity_images_folder}{The location of the folder containing the singularity images.
By default, this will use either the \code{DYNWRAP_SINGULARITY_IMAGES_FOLDER} environment variable,
the \code{dynwrap_singularity_images_folder} option, or otherwise the working directory (not recommended).}
}
\description{
These functions create a TI method from a docker image. Supports both docker and singularity as a backend.
}
