% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_create_ti_method_with_container.R
\name{create_ti_method_with_container}
\alias{create_ti_method_with_container}
\title{Create a TI method from a docker image}
\usage{
create_ti_method_with_container(image, config = container_config())
}
\arguments{
\item{image}{The name of the docker repository (e.g. \code{"dynverse/angle"}).
It is recommended to include a specific version (e.g. \code{"dynverse/angle@sha256:473e54..."}).}

\item{config}{A container config. See \code{\link[=container_config]{container_config()}} for more information.}
}
\description{
These functions create a TI method from a docker image. Supports both docker and singularity as a backend.
}
