% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_process_definition.R
\name{.method_process_definition}
\alias{.method_process_definition}
\title{Method process definition}
\usage{
.method_process_definition(definition, return_function)
}
\arguments{
\item{definition}{A definition, see \code{\link[=definition]{definition()}}}

\item{return_function}{Whether to return a function that allows you to override the default parameters, or just return the method meta data as is.}
}
\description{
Method process definition
}
\keyword{create_ti_method}
