% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_prior_information.R
\name{generate_prior_information}
\alias{generate_prior_information}
\title{Extract the prior information from the milestone network}
\usage{
generate_prior_information(cell_ids, milestone_ids, milestone_network,
  milestone_percentages, progressions, divergence_regions, expression,
  feature_info = NULL, cell_info = NULL, marker_fdr = 0.005,
  given = NULL)
}
\arguments{
\item{cell_ids}{The ids of the cells.}

\item{milestone_ids}{The ids of the milestones in the trajectory. Type: Character vector.}

\item{milestone_network}{The network of the milestones.
Type: Data frame(from = character, to = character, length = numeric, directed = logical).}

\item{milestone_percentages}{A data frame specifying what percentage milestone each cell
consists of.
Type: Data frame(cell_id = character, milestone_id = character, percentage = numeric).}

\item{progressions}{Specifies the progression of a cell along a transition in the milestone_network.
Type: Data frame(cell_id = character, from = character, to = character, percentage = numeric).}

\item{divergence_regions}{A data frame specifying the divergence
regions between milestones (e.g. a bifurcation).
Type: Data frame(divergence_id = character, milestone_id = character, is_start = logical).}

\item{expression}{The normalised expression values with genes in columns and cells in rows}

\item{feature_info}{Optional meta-information of the features, a data.frame with at least feature_id as column}

\item{cell_info}{Optional meta-information pertaining the cells.}

\item{marker_fdr}{Maximal FDR value for a gene to be considered a marker}

\item{given}{Prior information already calculated}
}
\description{
For example, what are the start cells, the end cells, to which milestone does each cell belong to.
}
