% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_process_definition.R
\name{get_default_parameters}
\alias{get_default_parameters}
\title{Get the default parameters of a method}
\usage{
get_default_parameters(definition)
}
\arguments{
\item{definition}{A TI method description}
}
\description{
Get the default parameters of a method
}
\keyword{create_ti_method}
