% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_infer_trajectory.R
\name{get_ti_methods}
\alias{get_ti_methods}
\title{Return all TI ti_methods}
\usage{
get_ti_methods(method_ids = NULL, as_tibble = TRUE,
  ti_packages = ifelse("dynmethods" \%in\%
  rownames(utils::installed.packages()), "dynmethods", "dynwrap"),
  evaluate = FALSE)
}
\arguments{
\item{method_ids}{The method identifiers. NULL if listing all methods}

\item{as_tibble}{Whether or not to return the ti_methods as a tibble}

\item{ti_packages}{In which packages to look for ti methods}

\item{evaluate}{Automatically evaluate the functions}
}
\description{
Return all TI ti_methods
}
