% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_infer_trajectory.R
\name{get_ti_methods}
\alias{get_ti_methods}
\title{Return all TI ti_methods}
\usage{
get_ti_methods(as_tibble = TRUE, packages = c("dynwrap"))
}
\arguments{
\item{as_tibble}{Whether or not to return the ti_methods as a tibble}

\item{packages}{In which packages to look for ti methods}
}
\description{
Return all TI ti_methods
}
