% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_infer_trajectory.R
\name{infer_trajectories}
\alias{infer_trajectories}
\alias{infer_trajectory}
\title{Infer trajectories}
\usage{
infer_trajectories(dataset, method, parameters = NULL,
  give_priors = NULL, seed = 1, mc_cores = 1, verbose = FALSE,
  capture_output = FALSE)

infer_trajectory(dataset, method, parameters = list(),
  give_priors = NULL, mc_cores = 1, verbose = FALSE, seed = 1, ...)
}
\arguments{
\item{dataset}{One or more datasets, as created using dynwrap}

\item{method}{One or more methods. Must be one of:
\itemize{
\item{an object or list of ti_... objects (eg. \code{dynmethods::ti_comp1()}),}
\item{a character vector containing the names of methods to execute (e.g. \code{"scorpius"}),}
\item{a character vector containing dockerhub repositories (e.g. \code{dynverse/paga}), or}
\item{a dynguidelines data frame.}
}}

\item{parameters}{A set of parameters to be used during trajectory inference.
A parameter set must be a named list of parameters.
If multiple methods were provided in the \code{method} parameter,
\code{parameters} must be an unnamed list of the same length.}

\item{give_priors}{All the priors a method is allowed to receive.
Must be a subset of all available priors (\code{\link[dynwrap:priors]{priors}}).}

\item{seed}{A seed to be set, if the method allows for it.}

\item{mc_cores}{The number of cores to use, allowing to parallellise the different datasets}

\item{verbose}{Whether or not to print information output}

\item{capture_output}{Whether to capture the stdout and stderr produced by a method}

\item{...}{Any additional parameters given to the method, will be concatednated to the parameters argument}
}
\description{
Infer trajectories
}
