% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_label_milestones.R
\name{label_milestones}
\alias{label_milestones}
\title{Label milestones based on highest expression of a set of markers}
\usage{
label_milestones(traj, labelling, expression_source = "expression",
  n_nearest_cells = 20)
}
\arguments{
\item{traj}{The trajectory}

\item{labelling}{List containing for each new name of a milestone the genes which will be used as markers}

\item{expression_source}{The expression source}
}
\description{
Label milestones based on highest expression of a set of markers
}
