% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_waypoints.R
\name{project_trajectory}
\alias{project_trajectory}
\title{Project a trajectory onto a dimensionality reduction using waypoints}
\usage{
project_trajectory(trajectory, dimred,
  waypoints = select_waypoints(trajectory))
}
\arguments{
\item{trajectory}{The trajectory as created by \code{\link[=infer_trajectory]{infer_trajectory()}} or \code{\link[=add_trajectory]{add_trajectory()}}}

\item{dimred}{The dimensionality reduction matrix (with cell_ids as rownames) or function which will run the dimensionality reduction}

\item{waypoints}{A list containing:
\itemize{
\item \code{waypoints}: a dataframe containing in the very least the waypoint_id
\item \code{milestone_percentages}: the positions of waypoints withing the trajectory
\item \code{geodesic_distances}: matrix with precalculated geodesic distances between waypoints (rows) and cells (columns), optional
A set of suitable waypoints can be created \code{\link[dynwrap:select_waypoints]{dynwrap::select_waypoints()}}
}}
}
\value{
A list containing dimred_segment_points and dimred_segment_progressions, which can be given to \code{\link[=add_dimred]{add_dimred()}}
}
\description{
Project a trajectory onto a dimensionality reduction using waypoints
}
