% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root_trajectory.R
\name{root_trajectory}
\alias{root_trajectory}
\title{Root trajectory}
\usage{
root_trajectory(trajectory, start_cell_id = NULL,
  start_milestone_id = trajectory$milestone_percentages \%>\% filter(cell_id
  == start_cell_id) \%>\% filter(percentage == max(percentage)) \%>\%
  pull(milestone_id))
}
\arguments{
\item{trajectory}{the trajectory object}

\item{start_cell_id}{The start cell id, not required if start_milestone_id is given}

\item{start_milestone_id}{The start milestone id, not required if start_cell_id is given}
}
\description{
Roots the trajectory by changing the directionality of all edges given a start cell
}
