% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_igraph_network.R
\name{simplify_trajectory}
\alias{simplify_trajectory}
\title{Simplify a trajectory}
\usage{
simplify_trajectory(traj)
}
\arguments{
\item{traj}{A trajectory to simplify}
}
\description{
Simplify a trajectory
}
