% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_trajectory.R
\name{simplify_trajectory}
\alias{simplify_trajectory}
\title{Simplify a trajectory by removing transient milestones}
\usage{
simplify_trajectory(traj, allow_self_loops = FALSE)
}
\arguments{
\item{traj}{A trajectory to simplify}

\item{allow_self_loops}{Whether or not to allow self loops.}
}
\description{
Simplify a trajectory by removing transient milestones
}
\keyword{adapt_trajectory}
