% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_trajectory.R
\name{simplify_trajectory}
\alias{simplify_trajectory}
\title{Simplify a trajectory by removing transient milestones}
\usage{
simplify_trajectory(trajectory, allow_self_loops = FALSE)
}
\arguments{
\item{trajectory}{The trajectory as created by \code{\link[=infer_trajectory]{infer_trajectory()}} or \code{\link[=add_trajectory]{add_trajectory()}}}

\item{allow_self_loops}{Whether or not to allow self loops.}
}
\description{
\itemize{
\item Milestones that are not a leaf or a branching point are removed: A -> B -> C becomes A -> C
\item Cycles contain at least 3 nodes, ie. A -> B -> A becomes A -> B -> C -> A
\item Loops are converted to a cycle, unless \code{allow_self_loops = TRUE}
}
}
\details{
The positions of the cells within the trajectory remain the same.
}
\keyword{adapt_trajectory}
