% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_trajectory_type.R
\name{simplify_trajectory_type}
\alias{simplify_trajectory_type}
\title{Convert directed trajectory type to simplified versions}
\usage{
simplify_trajectory_type(...)
}
\arguments{
\item{...}{trajectory types to simplify}
}
\description{
Convert directed trajectory type to simplified versions
}
\examples{
simplify_trajectory_type("directed_linear", "directed_graph", "binary_tree")
}
