% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_expression_to_wrapper.R
\name{wrap_expression}
\alias{wrap_expression}
\title{Wrap expression}
\usage{
wrap_expression(expression, counts, cell_info = NULL, feature_info = NULL,
  ..., id = "")
}
\arguments{
\item{expression}{the normalised expression values with genes in columns and cells in rows}

\item{counts}{The counts with genes in columns and cells in rows}

\item{cell_info}{Optional meta-information pertaining the cells.}

\item{feature_info}{Optional meta-information pertaining the features}

\item{...}{Extra information to be stored in the wrapper.}

\item{id}{A unique identifier for the data}
}
\description{
Wrap expression
}
