% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_output.R
\name{wrap_output}
\alias{wrap_output}
\alias{wrap_rds}
\alias{wrap_text}
\alias{wrap_feather}
\title{Wrap the output of a TI method}
\usage{
wrap_output(cell_ids, output_ids, dir_output, output_format = c("text",
  "rds", "feather", "dynwrap"))

wrap_rds(cell_ids, output_ids, dir_output)

wrap_text(cell_ids, output_ids, dir_output)

wrap_feather(cell_ids, output_ids, dir_output)
}
\arguments{
\item{cell_ids}{The cell_ids of the dataset}

\item{output_ids}{The ids of the outputs generated by the methods}

\item{dir_output}{The directory containing the output files}

\item{output_format}{The output format, can be rds, text or dynwrap}
}
\description{
Will extract outputs from a given folder given an output format:
\itemize{
\item rds: One \code{output.rds} file containing all the output. Output can be present as \code{.$pseudotime} or as \code{.$linear_trajectory$pseudotime} . Parameters specific for an output should be given in \code{.$linear_trajectory$params}
\item text: Csv and/or json files, including subdirectories. Output can be present as \code{./pseudotime.csv} or as \code{./linear_trajectory/pseudotime.csv}. Parameters specific for an output should be given in \code{./linear_trajectory/params.json}.
\item feather: Feather files
\item dynwrap: Directly a dynwrap wrapper as an rds file in \code{output.rds}
}
}
