% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_output.R, R/wrap_output_dynwrap.R,
%   R/wrap_output_rds.R, R/wrap_output_text.R
\name{wrap_output}
\alias{wrap_output}
\alias{wrap_dynwrap}
\alias{wrap_rds}
\alias{wrap_text}
\title{Wrap the output of a TI method}
\usage{
wrap_output(output_ids, dir_output, output_format = c("text", "rds",
  "dynwrap"))

wrap_dynwrap(output_ids, dir_output)

wrap_rds(output_ids, dir_output)

wrap_text(output_ids, dir_output)
}
\arguments{
\item{output_ids}{The outputs ids expected by the method.}

\item{dir_output}{The directory containing the output files.}

\item{output_format}{The output format. Must be one of \code{"text"}, \code{"rds"} or \code{"dynwrap"}.}
}
\description{
Will extract outputs from a given folder given an output format:
\itemize{
\item rds: One \code{output.rds} file containing all the output.\
\itemize{
\item Output can be present as \code{.$pseudotime} or as \code{.$linear_trajectory$pseudotime}.
\item Parameters specific for an output should be given in \code{.$linear_trajectory$params}
}
\item text: Csv and/or json files, including subdirectories.
\itemize{
\item Output can be present as \code{./pseudotime.csv} or as \code{./linear_trajectory/pseudotime.csv}.
\item Parameters specific for an output should be given in \code{./linear_trajectory/params.json}.
}
\item dynwrap: Directly a dynwrap wrapper as an rds file in \code{output.rds}
}
}
